Testsuite: T861: use proper base MAC address from RFC7042
Commit 085df7615a ("Testsuite: T861: always use 2 VCPUs") also altered the base
MAC address used by QEMU to a locally administered one. Something that looked
"right" in the beginning turned out to break the smoketest platform.

The reason is the locally administered bit is evaluated in [1] and if set and
not on the exclusion list (as it was a Realtek base MAC address before), the
interface in question is not considered persistent and thus not added to the
configuration file upon system startup.

1: https://github.com/vyos/vyos-1x/blob/825743b6bcdf8fa2c263dabaa3fee40ba7a98525/src/helpers/vyos-interface-rescan.py#L73-L74